+++ features/support/config.rb
@@ -15,6 +15,6 @@ def set_extract_args args
     @extract_args = args
 end
 
-def set_prepare_args args
-    @prepare_args = args
+def set_contract_args args
+    @contract_args = args
 end
